feat: Connect yt url fetching.
