--- utils/utils.py
@@ -1,19 +1,0 @@
-from langchain_community.document_loaders import PyPDFLoader
-from langchain.text_splitter import RecursiveCharacterTextSplitter
-from langchain_community.document_loaders.text import TextLoader
-
-
-def get_documents_txt(path):
-    loader = TextLoader(path, encoding='utf-8')
-    pages = loader.load_and_split()
-
-    text_splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=10)
-    return text_splitter.split_documents(pages)
-
-
-def get_documents_pdf(path):
-    loader = PyPDFLoader(path)
-    pages = loader.load_and_split()
-
-    text_splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=10)
-    return text_splitter.split_documents(pages)
